HSEARCH-3717 Move options common to all index field types to a dedicated interface and abstract class
