+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/types/dsl/ElasticsearchNativeIndexFieldTypeOptionsStep.java
@@ -6,10 +6,9 @@
  */
 package org.hibernate.search.backend.elasticsearch.types.dsl;
 
-import org.hibernate.search.engine.backend.types.dsl.IndexFieldTypeConverterStep;
-import org.hibernate.search.engine.backend.types.dsl.IndexFieldTypeFinalStep;
+import org.hibernate.search.engine.backend.types.dsl.IndexFieldTypeOptionsStep;
 
 public interface ElasticsearchNativeIndexFieldTypeOptionsStep<S extends ElasticsearchNativeIndexFieldTypeOptionsStep<?>>
-		extends IndexFieldTypeConverterStep<S, String>, IndexFieldTypeFinalStep<String> {
+		extends IndexFieldTypeOptionsStep<S, String> {
 
 }
